Moved can_draw_ui() for general use
