+++ menu/menu_body.py
@@ -3,7 +3,7 @@ import bpy
 from ..misc.prop_utils import evaluate_rna
 from ..misc.ui_multiline import label_multiline
 from ..model_selection.active_object import mustardui_active_object
-from ..warnings.ops_fix_old_UI import can_draw_ui
+from ..warnings.can_draw_ui import can_draw_ui
 from . import MainPanel
