+++ menu/menu_hair.py
@@ -3,7 +3,7 @@ import bpy
 from .. import __package__ as base_package
 from ..configuration.naming_convention import strip_naming_convention
 from ..model_selection.active_object import mustardui_active_object
-from ..warnings.ops_fix_old_UI import can_draw_ui
+from ..warnings.can_draw_ui import can_draw_ui
 from . import MainPanel
 from .misc import mustardui_custom_properties_print
