+++ menu/menu_select_model.py
@@ -3,7 +3,7 @@ from collections import Counter
 import bpy
 
 from ..model_selection.active_object import mustardui_active_object
-from ..warnings.ops_fix_old_UI import can_draw_ui
+from ..warnings.can_draw_ui import can_draw_ui
 from . import MainPanel
