+++ menu/menu_settings.py
@@ -1,7 +1,7 @@
 import bpy
 
 from ..model_selection.active_object import mustardui_active_object
-from ..warnings.ops_fix_old_UI import can_draw_ui
+from ..warnings.can_draw_ui import can_draw_ui
 from . import MainPanel
