+++ warnings/ops_fix_old_UI.py
@@ -1,16 +1,7 @@
 import bpy
 from bpy.props import BoolProperty
 
-
-# Checks if there are some conditions blocking the menus to be drawn
-def can_draw_ui():
-    import bpy
-
-    # Check presence of old UI scripts
-    for file in bpy.data.texts:
-        if "mustard_ui.py" in file.name:
-            return True
-    return False
+from .can_draw_ui import can_draw_ui
 
 
 class MustardUI_Warnings_FixOldUI(bpy.types.Operator):
